requirements.txt
numpy==1.23.5
pandas==1.5.3
torch==2.0.1
pytorch-lightning==1.9.4
pytorch-forecasting==1.0.0
datasets==3.1.0
tensorboard==2.14.0
pygrinder==0.7
h5py==3.11.0
tsdb==0.7.1
einops==0.8.1
sympy==1.13.3
